Recreate migrations, add extra payment validation

+++ mortgage/api/views.py
@@ -1,3 +1,4 @@
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.generics import RetrieveUpdateDestroyAPIView, ListCreateAPIView, ListAPIView
 from rest_framework.permissions import IsAuthenticated
@@ -5,7 +6,7 @@ from rest_framework.permissions import IsAuthenticated
 from datetime import datetime
 
 from mortgage.models import Mortgage, Payment
-from .serializers import MortgageSerializer, PaymentSerializer
+from .serializers import BasicPaymentSerializer, MortgageSerializer, PaymentSerializer
 from .permissions import IsOwner
 from .pagination import CustomPagination
 from .payment_schedule import PaymentScheduler
@@ -52,7 +53,11 @@ class CalcPaymentsSchedule(ListAPIView):
     pagination_class = CustomPagination
 
     def post(self, request, *args, **kwargs):
-        current_mortgage = Mortgage.git_mortgage(kwargs['mortgage_id'])
+        mortgage_id = kwargs['mortgage_id']
+        current_mortgage = Mortgage.get_mortgage(mortgage_id)
+        if not current_mortgage:
+            return Response(data={'error': Mortgage.get_not_found_error_message(mortgage_id)},
+                            status=status.HTTP_404_NOT_FOUND)
 
         current_payments = Payment.objects.filter(mortgage_id=current_mortgage.id)
         current_payments.delete()  # delete old payment schedule if it exists
@@ -64,15 +69,25 @@ class CalcPaymentsSchedule(ListAPIView):
 
 
 class AddExtraPayment(ListAPIView):
-    serializer_class = PaymentSerializer
+    serializer_class = BasicPaymentSerializer
     queryset = Payment.objects.all()
     permission_classes = [IsAuthenticated]
     pagination_class = CustomPagination
 
     def post(self, request, *args, **kwargs):
-        current_mortgage = Mortgage.git_mortgage(kwargs['mortgage_id'])
+        mortgage_id = kwargs['mortgage_id']
+        current_mortgage = Mortgage.get_mortgage(mortgage_id)
+        if not current_mortgage:
+            return Response(data={'error': Mortgage.get_not_found_error_message(mortgage_id)},
+                            status=status.HTTP_404_NOT_FOUND)
         current_payments = Payment.objects.filter(mortgage_id=current_mortgage.id)
 
+        serializer = self.serializer_class(data=request.data)
+        if not serializer.is_valid():
+            return Response(
+                serializer.errors,
+                status=status.HTTP_400_BAD_REQUEST
+            )
         extra_payment = Payment(
             mortgage=current_mortgage,
             amount=request.data['amount'],
